+++ configure
@@ -354,13 +354,13 @@ PREFIX=${PREFIX:-/usr/local}
 REMOTE_PREFIX=${REMOTE_PREFIX:-/usr/local}
 LIB_DIR=${LIB_DIR:-${PREFIX}/lib}
 REMOTE_LIB_DIR=${REMOTE_LIB_DIR:-${REMOTE_PREFIX}/lib}
-HEADER_DIR=${HEADER_DIR:-${PREFIX}/include/RF24}
-REMOTE_HEADER_DIR=${REMOTE_HEADER_DIR:-${REMOTE_PREFIX}/include/RF24}
+HEADER_DIR=${HEADER_DIR:-${PREFIX}/include/RF24_c}
+REMOTE_HEADER_DIR=${REMOTE_HEADER_DIR:-${REMOTE_PREFIX}/include/RF24_c}
 EXAMPLES_DIR=${EXAMPLES_DIR:-${PREFIX}/bin}
 REMOTE_EXAMPLES_DIR=${REMOTE_EXAMPLES_DIR:-${REMOTE_PREFIX}/bin}
 LDCONFIG=${LDCONFIG-ldconfig}
 REMOTE_LDCONFIG=${REMOTE_LDCONFIG-/sbin/ldconfig}
-LIB=${LIB:-rf24}
+LIB=${LIB:-rf24_c}
 LIB_VERSION=${LIB_VERSION:-$(awk -F "=" '/version/ {print $2}' library.properties)}
 LIB_DEPRECATE_NAME=${LIB_DEPRECATE_NAME:-"rf24-bcm"}
 LIB_DEPRECATE_VERSION=${LIB_DEPRECATE_VERSION:-""}
