+++ library.properties
@@ -1,4 +1,4 @@
-name=RF24
+name=RF24_c
 version=1.2.0
 author=TMRh20
 maintainer=TMRh20
